split vim into core & cmd map
