+++ configs/configs.lua
@@ -1,0 +1,13 @@
+local json = require("hs.json")
+local configsJsonPath
+
+local function loadConfig()
+    local file = io.open(configsJsonPath,"r")
+    if not file then return {} end
+    local content = file:read("*a")
+    file:close()
+    return json.decode(content)
+end
+
+local configs = loadConfig()
+return configs
